move stuff to a v3_migration package
