+++ dojo/api_v2/views.py
@@ -3192,322 +3192,3 @@ class NotificationWebhooksViewSet(
     filter_backends = (DjangoFilterBackend,)
     filterset_fields = "__all__"
     permission_classes = (permissions.IsSuperUser, DjangoModelPermissions)  # TODO: add permission also for other users
-
-
-# V3
-
-from django_filters.rest_framework import FilterSet
-class AssetApiScanConfigurationFilterSet(FilterSet):
-    pass
-
-# Authorization: object-based
-@extend_schema_view(**schema_with_prefetch())
-class AssetAPIScanConfigurationViewSet(
-    PrefetchDojoModelViewSet,
-):
-    serializer_class = serializers.AssetAPIScanConfigurationSerializer
-    queryset = Product_API_Scan_Configuration.objects.none()
-    filter_backends = (DjangoFilterBackend,)
-    filterset_fields = [
-        "id",
-        "product",
-        "tool_configuration",
-        "service_key_1",
-        "service_key_2",
-        "service_key_3",
-    ]
-    permission_classes = (
-        IsAuthenticated,
-        permissions.UserHasProductAPIScanConfigurationPermission,
-    )
-
-    def get_queryset(self):
-        return get_authorized_product_api_scan_configurations(
-            Permissions.Product_API_Scan_Configuration_View,
-        )
-
-
-@extend_schema_view(**schema_with_prefetch())
-class AssetViewSet(
-    prefetch.PrefetchListMixin,
-    prefetch.PrefetchRetrieveMixin,
-    mixins.CreateModelMixin,
-    mixins.DestroyModelMixin,
-    mixins.UpdateModelMixin,
-    viewsets.GenericViewSet,
-    dojo_mixins.DeletePreviewModelMixin,
-):
-    serializer_class = serializers.AssetSerializer
-    queryset = Product.objects.none()
-    filter_backends = (DjangoFilterBackend,)
-    filterset_class = ApiProductFilter
-    permission_classes = (
-        IsAuthenticated,
-        permissions.UserHasProductPermission,
-    )
-
-    def get_queryset(self):
-        return get_authorized_products(Permissions.Product_View).distinct()
-
-    def destroy(self, request, *args, **kwargs):
-        instance = self.get_object()
-        if get_setting("ASYNC_OBJECT_DELETE"):
-            async_del = async_delete()
-            async_del.delete(instance)
-        else:
-            instance.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-    # def list(self, request):
-    #     # Note the use of `get_queryset()` instead of `self.queryset`
-    #     queryset = self.get_queryset()
-    #     serializer = self.serializer_class(queryset, many=True)
-    #     return Response(serializer.data)
-
-    @extend_schema(
-        request=serializers.ReportGenerateOptionSerializer,
-        responses={status.HTTP_200_OK: serializers.ReportGenerateSerializer},
-    )
-    @action(
-        detail=True, methods=["post"], permission_classes=[IsAuthenticated],
-    )
-    def generate_report(self, request, pk=None):
-        product = self.get_object()
-
-        options = {}
-        # prepare post data
-        report_options = serializers.ReportGenerateOptionSerializer(
-            data=request.data,
-        )
-        if report_options.is_valid():
-            options["include_finding_notes"] = report_options.validated_data[
-                "include_finding_notes"
-            ]
-            options["include_finding_images"] = report_options.validated_data[
-                "include_finding_images"
-            ]
-            options[
-                "include_executive_summary"
-            ] = report_options.validated_data["include_executive_summary"]
-            options[
-                "include_table_of_contents"
-            ] = report_options.validated_data["include_table_of_contents"]
-        else:
-            return Response(
-                report_options.errors, status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        data = report_generate(request, product, options)
-        report = serializers.ReportGenerateSerializer(data)
-        return Response(report.data)
-
-
-# Authorization: object-based
-@extend_schema_view(**schema_with_prefetch())
-class AssetMemberViewSet(
-    PrefetchDojoModelViewSet,
-):
-    serializer_class = serializers.AssetMemberSerializer
-    queryset = Product_Member.objects.none()
-    filter_backends = (DjangoFilterBackend,)
-    filterset_fields = ["id", "product_id", "user_id"]
-    permission_classes = (
-        IsAuthenticated,
-        permissions.UserHasProductMemberPermission,
-    )
-
-    def get_queryset(self):
-        return get_authorized_product_members(
-            Permissions.Product_View,
-        ).distinct()
-
-    @extend_schema(
-        exclude=True,
-    )
-    def partial_update(self, request, pk=None):
-        # Object authorization won't work if not all data is provided
-        response = {"message": "Patch function is not offered in this path."}
-        return Response(response, status=status.HTTP_405_METHOD_NOT_ALLOWED)
-
-
-# Authorization: object-based
-@extend_schema_view(**schema_with_prefetch())
-class AssetGroupViewSet(
-    PrefetchDojoModelViewSet,
-):
-    serializer_class = serializers.AssetGroupSerializer
-    queryset = Product_Group.objects.none()
-    filter_backends = (DjangoFilterBackend,)
-    filterset_fields = ["id", "product_id", "group_id"]
-    permission_classes = (
-        IsAuthenticated,
-        permissions.UserHasProductGroupPermission,
-    )
-
-    def get_queryset(self):
-        return get_authorized_product_groups(
-            Permissions.Product_Group_View,
-        ).distinct()
-
-    @extend_schema(
-        exclude=True,
-    )
-    def partial_update(self, request, pk=None):
-        # Object authorization won't work if not all data is provided
-        response = {"message": "Patch function is not offered in this path."}
-        return Response(response, status=status.HTTP_405_METHOD_NOT_ALLOWED)
-
-
-# Authorization: object-based
-@extend_schema_view(**schema_with_prefetch())
-class OrganizationViewSet(
-    PrefetchDojoModelViewSet,
-):
-    serializer_class = serializers.OrganizationSerializer
-    queryset = Product_Type.objects.none()
-    filter_backends = (DjangoFilterBackend,)
-    filterset_fields = [
-        "id",
-        "name",
-        "critical_product",
-        "key_product",
-        "created",
-        "updated",
-    ]
-    permission_classes = (
-        IsAuthenticated,
-        permissions.UserHasProductTypePermission,
-    )
-
-    def get_queryset(self):
-        return get_authorized_product_types(
-            Permissions.Product_Type_View,
-        ).distinct()
-
-    # Overwrite perfom_create of CreateModelMixin to add current user as owner
-    def perform_create(self, serializer):
-        serializer.save()
-        product_type_data = serializer.data
-        product_type_data.pop("authorization_groups")
-        product_type_data.pop("members")
-        member = Product_Type_Member()
-        member.user = self.request.user
-        member.product_type = Product_Type(**product_type_data)
-        member.role = Role.objects.get(is_owner=True)
-        member.save()
-
-    def destroy(self, request, *args, **kwargs):
-        instance = self.get_object()
-        if get_setting("ASYNC_OBJECT_DELETE"):
-            async_del = async_delete()
-            async_del.delete(instance)
-        else:
-            instance.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-    @extend_schema(
-        request=serializers.ReportGenerateOptionSerializer,
-        responses={status.HTTP_200_OK: serializers.ReportGenerateSerializer},
-    )
-    @action(
-        detail=True, methods=["post"], permission_classes=[IsAuthenticated],
-    )
-    def generate_report(self, request, pk=None):
-        product_type = self.get_object()
-
-        options = {}
-        # prepare post data
-        report_options = serializers.ReportGenerateOptionSerializer(
-            data=request.data,
-        )
-        if report_options.is_valid():
-            options["include_finding_notes"] = report_options.validated_data[
-                "include_finding_notes"
-            ]
-            options["include_finding_images"] = report_options.validated_data[
-                "include_finding_images"
-            ]
-            options[
-                "include_executive_summary"
-            ] = report_options.validated_data["include_executive_summary"]
-            options[
-                "include_table_of_contents"
-            ] = report_options.validated_data["include_table_of_contents"]
-        else:
-            return Response(
-                report_options.errors, status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        data = report_generate(request, product_type, options)
-        report = serializers.ReportGenerateSerializer(data)
-        return Response(report.data)
-
-
-# Authorization: object-based
-@extend_schema_view(**schema_with_prefetch())
-class OrganizationMemberViewSet(
-    PrefetchDojoModelViewSet,
-):
-    serializer_class = serializers.OrganizationMemberSerializer
-    queryset = Product_Type_Member.objects.none()
-    filter_backends = (DjangoFilterBackend,)
-    filterset_fields = ["id", "product_type_id", "user_id"]
-    permission_classes = (
-        IsAuthenticated,
-        permissions.UserHasProductTypeMemberPermission,
-    )
-
-    def get_queryset(self):
-        return get_authorized_product_type_members(
-            Permissions.Product_Type_View,
-        ).distinct()
-
-    def destroy(self, request, *args, **kwargs):
-        instance = self.get_object()
-        if instance.role.is_owner:
-            owners = Product_Type_Member.objects.filter(
-                product_type=instance.product_type, role__is_owner=True,
-            ).count()
-            if owners <= 1:
-                return Response(
-                    "There must be at least one owner",
-                    status=status.HTTP_400_BAD_REQUEST,
-                )
-        self.perform_destroy(instance)
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-    @extend_schema(
-        exclude=True,
-    )
-    def partial_update(self, request, pk=None):
-        # Object authorization won't work if not all data is provided
-        response = {"message": "Patch function is not offered in this path."}
-        return Response(response, status=status.HTTP_405_METHOD_NOT_ALLOWED)
-
-
-# Authorization: object-based
-@extend_schema_view(**schema_with_prefetch())
-class OrganizationGroupViewSet(
-    PrefetchDojoModelViewSet,
-):
-    serializer_class = serializers.OrganizationGroupSerializer
-    queryset = Product_Type_Group.objects.none()
-    filter_backends = (DjangoFilterBackend,)
-    filterset_fields = ["id", "product_type_id", "group_id"]
-    permission_classes = (
-        IsAuthenticated,
-        permissions.UserHasProductTypeGroupPermission,
-    )
-
-    def get_queryset(self):
-        return get_authorized_product_type_groups(
-            Permissions.Product_Type_Group_View,
-        ).distinct()
-
-    @extend_schema(
-        exclude=True,
-    )
-    def partial_update(self, request, pk=None):
-        # Object authorization won't work if not all data is provided
-        response = {"message": "Patch function is not offered in this path."}
-        return Response(response, status=status.HTTP_405_METHOD_NOT_ALLOWED)
