+++ dojo/v3_migration/__init__.py
@@ -1,0 +1,4 @@
+from dojo.v3_migration.utils import (
+    v3_migration_enabled,
+    redirect_view,
+)
